Main window made
The main window, with all the options and functionalities, is designed and implemented. When the user make the login, the program shows the main window with the icons of the main functionalities.
Also, there are two more exceptions that verify if the user filled all the info of a new user or a try to login.